Refactor panels to views
